+++ Sketchbook/libraries/M5Unit-Encoder/src/UNIT_ENCODER.h
@@ -3,7 +3,7 @@
  * @copyright Copyright (c) 2022 by M5Stack[https://m5stack.com]
  *
  * @Links [Unit Encoder](https://docs.m5stack.com/en/unit/encoder)
- * @version  V0.0.1
+ * @version  V0.0.2
  * @date  2022-07-11
  */
 #ifndef _UNIT_ENCODER_H_
@@ -24,7 +24,7 @@ class Unit_Encoder {
     TwoWire* _wire;
     uint8_t _scl;
     uint8_t _sda;
-    uint8_t _speed;
+    uint32_t _speed;
     void writeBytes(uint8_t addr, uint8_t reg, uint8_t* buffer, uint8_t length);
     void readBytes(uint8_t addr, uint8_t reg, uint8_t* buffer, uint8_t length);
